src/conversation/requirements.txt
supabase==2.3.4
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0
asyncio==3.4.3
aiohttp==3.9.1
faker==22.0.0
psycopg2-binary==2.9.9
croniter==2.0.1
